model-invocation-endpoint/requirements.txt
mlflow
sklearn
pandas
numpy
fastapi
uvicorn[standard]
python-dotenv